Add joinUrl in uriParser utility
